+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 * The UUID is no longer required when adding a dependency.
+* If the project is explicitly specified, it is considered "shared" and existing
+  dependencies are never removed.
 
 ## v0.1.19 (2025-09-17)
 * Add the CLI.